Include toodoo package
This package introduces the `TodoList` collection type along with
convenience methods
