+++ toodoo/todo.go
@@ -1,0 +1,6 @@
+package toodoo
+
+type Todo struct {
+	Name     string `json:"name"`
+	Complete bool   `json:"is_complete"`
+}
